linkedin-agent-backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.28.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching and task queue
redis==5.0.1
celery==5.3.4
flower==2.0.1

# AI/ML
langchain==0.0.335
langchain-openai==0.0.2
langchain-anthropic==0.1.1
openai==1.3.5
anthropic==0.7.4
pinecone-client==2.2.4
tiktoken==0.5.1

# LinkedIn API
requests==2.31.0
oauthlib==3.2.2
requests-oauthlib==1.3.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
httpx==0.25.1
pandas==2.1.3
numpy==1.26.2
pyyaml==6.0.1
jinja2==3.1.2

# Logging and observability
structlog==23.2.0
prometheus-client==0.17.1
sentry-sdk==1.32.0
ddtrace==1.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Linting and formatting
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0 